_old.py
import openai
import chromadb
import os
from dotenv import load_dotenv

class EmbeddingsModel:
    def __init__(self, api_key):
        openai.api_key = api_key

    def generate_embeddings(self, file_path):
        # Code to read the file and generate embeddings using OpenAI APIs
        embeddings = openai.generate_embeddings(file_path)
        
        # Code to store the embeddings into a Chroma DB in a named collection
        chromadb.store_embeddings(embeddings, collection_name)

class QueryModel:
    def __init__(self, api_key):
        openai.api_key = api_key

    def retrieve_context(self, collection_name):
        # Code to retrieve the context for the given collection from Chroma DB
        context = chromadb.retrieve_context(collection_name)
        
        return context

    def infer_query(self, user_query):
        # Code to infer the query using perplexity API
        query = openai.infer_perplexity(user_query)
        
        return query

    def provide_response(self, query):
        # Code to provide response using perplexity API
        response = openai.provide_response(query)
        
        return response

# Usage example
# Load environment variables from .env file
load_dotenv()

# Get the API key from environment variable
api_key = os.getenv("OPENAI_API_KEY")
infer_key = os.getenv("PPLEX_API_KEY")

embeddings_model = EmbeddingsModel(api_key)
embeddings_model.generate_embeddings("file_path")

query_model = QueryModel(api_key)
context = query_model.retrieve_context("collection_name")
query = query_model.infer_query("user_query")
response = query_model.provide_response(query)

embeddings_model = EmbeddingsModel(api_key)
embeddings_model.generate_embeddings("file_path")

query_model = QueryModel(api_key)
context = query_model.retrieve_context("collection_name")
query = query_model.infer_query("user_query")
response = query_model.provide_response(query)
